Current State of Anger Level
I clicked too many of the wrong buttons and ended up merging anger level with things it shouldn't be merged with. The result: spending 3 hours removing duplicate code and rewriting some portions until the level could be run. I couldn't get crumbling platforms to work, so that's something that'll look into tomorrow. Besides that, everything seems to be working like it used to.