Anger Level as of March 17 2021
Included an if statement in knockbacktrigger.cpp to check and see if the trigger is a spike. 
Turns out I was using outdated versions of the crumble json and png, which explained why they didn't work.
The moving platform was moving the player I adjusted the value of passiveMovement in player.cpp