Abigail can jump!
The shrinkX and shrinkY for Abigail's entity was off, which explains why Abigail couldn't jump.
Cleaned up the update function a bit.
The level is technically playable now, however the player would need to be holding the sprint key(shift) in order to make most of the jumps in the level.